[trainer] move shared PPO helpers out of ray trainer
