+++ tests/trainer/ppo/test_spec_decode_metrics_on_cpu.py
@@ -13,11 +13,8 @@
 # limitations under the License.
 
 import numpy as np
-import pytest
 
-pytest.importorskip("ray")
-
-from verl.trainer.ppo.ray_trainer import compute_spec_decode_metrics
+from verl.trainer.ppo.trainer_utils import compute_spec_decode_metrics
 
 
 def test_spec_decode_metrics_detect_drafts_with_zero_acceptance():
