+++ tests/workers/rollout/rollout_vllm/test_mtp_hybrid_sleep_acceptance_on_cpu.py
@@ -21,7 +21,7 @@ import pytest
 pytest.importorskip("ray")
 pytest.importorskip("vllm")
 
-from verl.trainer.ppo.ray_trainer import compute_spec_decode_metrics
+from verl.trainer.ppo.trainer_utils import compute_spec_decode_metrics
 from verl.workers.rollout.vllm_rollout import vllm_async_server
