+++ verl/trainer/ppo/v1/utils.py
@@ -19,7 +19,7 @@ import torch
 
 from verl.protocol import DataProto
 from verl.trainer.ppo import core_algos
-from verl.trainer.ppo.ray_trainer import compute_advantage
+from verl.trainer.ppo.trainer_utils import compute_advantage
 from verl.trainer.ppo.v1.replay_buffer import DAPO_FILTERED_REWARD_COUNTS_KEY
